// generated/src/app/api-models/custserv-code.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CustservCode {
  csCode: string;
  codeType: string;
  description: string;
  adMaintainFlag: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}